requirements.txt
amqp==2.5.2
asgiref==3.2.7
autopep8==1.5.2
billiard==3.6.3.0
celery==4.4.2
dj-database-url==0.5.0
Django==3.0.6
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-environ==0.4.5
django-extensions==2.2.9
django-timezone-field==4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
importlib-metadata==1.6.0
kombu==4.6.8
psycopg2-binary==2.8.5
pycodestyle==2.6.0
PyJWT==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.2
six==1.15.0
sqlparse==0.3.1
vine==1.3.0
zipp==3.1.0
